+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project follows [Semantic Versioning](https://semver.org/) and documents ch
 ### Changed
 
 ## [2.0.1]
+- `validate` task introduced to allow for validation of psfgen-produced PSF and PDB files
 - `include` and `exclude` now allow full logical expressions for better control of inclusion and exclusion of atoms and residues from source structures
 - `fetch` task now separately responsible for downloading necessary pdb files
 - `continuation` task now separately responsible for starting from a given state (PSF/PDB/COOR/XSC)
@@ -14,7 +15,6 @@ This project follows [Semantic Versioning](https://semver.org/) and documents ch
 
 ## [1.21.2] - 2025-07-17
 - first official Zenodo release [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.16051499.svg)](https://doi.org/10.5281/zenodo.16051499)
-
 - fixed bug in user-modifiable segtype handling
 
 ## [1.21.1] - 2025-07-15